% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_unidade.R
\name{correct_unidade}
\alias{correct_unidade}
\title{Correct the unidade}
\usage{
correct_unidade(unidade, unidades)
}
\arguments{
\item{unidade}{An unidade}

\item{unidades}{A vector of all correct unidades}
}
\value{
The correct unidade
}
\description{
Try to find the right unidade of a non-conforming unidade
}
