% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ihmt_presentation.R
\name{generate_ihmt_presentation}
\alias{generate_ihmt_presentation}
\title{Generate IHMT presentation}
\usage{
generate_ihmt_presentation(
  output_dir = NULL,
  output_file = "ihmt_presentation.pdf"
)
}
\arguments{
\item{output_dir}{The directory to which the file should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}

\item{date}{A date to be printed on the report. If \code{NULL} (the
default), the current date will be used}
}
\value{
An html will be written
}
\description{
Generate the presentation for the Oct 30-31 IHMT presentation
}
