% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_maragra_clinical_malaria_incidence.R
\name{generate_maragra_clinical_malaria_incidence}
\alias{generate_maragra_clinical_malaria_incidence}
\title{Generate Maragra clinical malaria incidence report}
\usage{
generate_maragra_clinical_malaria_incidence(date = NULL, output_dir = NULL,
  output_file = "malaria_clinical_malaria_incidence.html")
}
\arguments{
\item{date}{A date to be printed on the report. If \code{NULL} (the 
default), the current date will be used}

\item{output_dir}{The directory to which the file should be written. If 
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}
}
\value{
An html will be written
}
\description{
Generate an html report of Maragra clinical malaria incidence, based
on aggregate numbers manually collected by Joe in December 2016 at the 
Maragra clinic and laboratory
}

