% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preliminary_analysis.R
\name{generate_preliminary_analysis}
\alias{generate_preliminary_analysis}
\title{Generate preliminary analysis report}
\usage{
generate_preliminary_analysis(date = NULL, output_dir = NULL,
  output_file = "maragra_preliminary_analysis.html")
}
\arguments{
\item{date}{A date to be printed on the report. If \code{NULL} (the 
default), the current date will be used}

\item{output_dir}{The directory to which the file should be written. If 
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}
}
\value{
An html will be written
}
\description{
Generate an html report of the preliminary analysis report,
originally prepared in August / September 2017
}

