% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_authors.R
\name{make_authors}
\alias{make_authors}
\title{Make authors}
\usage{
make_authors(authors = data.frame(name = c("Joe Brew", "Kizito Gondo",
  "Elton Dorkin", "Eduardo Nhamahanga", "Menno Pradhan", "Laia Cirera",
  "Ranjeeta Thomas", "Elisa Sicuri"), email = c("joe@databrew.cc",
  "KGondo@illovo.co.za", "EDorkin@illovo.co.za", "ENhamahanga@illovo.co.za",
  "m.p.pradhan@vu.nl", "laia.cirera@isglobal.org",
  "ranjeeta.thomas@imperial.ac.uk", "elisa.sicuri@isglobal.org"), affilitation =
  c("isglobal,cism,vu", "ma", "ma", "ma", "vu,uva", "isglobal,cism", "icl",
  "isglobal,cism,icl"), footnote = c("Corresponding Author",      rep("", 7)),
  stringsAsFactors = FALSE), institutions = data.frame(code = c("isglobal",
  "icl", "cism", "vu", "uva", "ma"), institution =
  c("Barcelona Institute for Global Health", "Imperial College London",
  "Centro de Investigação em Saúde de Manhiça", "VU University Amsterdam",
  "University of Amsterdam",
  "Maragra Açucar SA, Subsidiary of Illovo Sugar Ltd"), country = c("Spain",
  "UK", "Mozambique", "Netherlands", "Netherlands", "Mozambique"), address =
  c("c/ Rosselló, 132, 5è 2a. 08036, Barcelona, Spain",
  "South Kensington Campus, London SW7 2AZ, U.K.",     
  "Vila da Manhiça, Bairro Cambeve, Rua 12, Distrito da Manhiça, CP 1929, Maputo, Mozambique",
  "De Boelelaan 1105, 1081 HV Amsterdam, Netherlands",
  "REC E, Roetersstraat 11, Amsterdam, Netherlands",
  "CP 2789, Maputo, Mozambique"), stringsAsFactors = FALSE), cat_it = TRUE,
  include_address = TRUE, include_country = TRUE, seperator = "\\n")
}
\arguments{
\item{authors}{A dataframe with columns: name, email, affiliation, footnote}

\item{institutions}{A dataframe with columns: code, institution, address}

\item{cat_it}{Whether to cat (as opposed to return)}

\item{include_address}{Whether to include the address}

\item{include_country}{Whether to include the country; only relevant if include_address is FALSE}

\item{seperator}{What to put between author names: usually a comma or line break}
}
\value{
Latex text
}
\description{
Cat the authors in latex-ready format
}
