% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_maragra.R
\name{theme_maragra}
\alias{theme_maragra}
\title{Theme for ggplots}
\usage{
theme_maragra(base_size = 14)
}
\arguments{
\item{base_size}{The base font size}
}
\value{
A ggplot compatible theme call
}
\description{
Add a theme to a ggplot object
}
