% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_analyses.R
\name{run_all_analyses}
\alias{run_all_analyses}
\title{Run all analyses}
\usage{
run_all_analyses(only = NULL, except = NULL, date = NULL,
  output_dir = NULL)
}
\arguments{
\item{only}{A vector of .Rmd file names (excluding path) to be knitted}

\item{except}{A vector or .Rmd file names (excluding path) to be excluded}

\item{date}{A date to be printed on the report. If \code{NULL} (the 
default), the current date will be used}

\item{output_dir}{The directory to which the analyses should be written. If 
\code{NULL} (the default), the current working directory will be used.}
}
\value{
html and pdf files will be written
}
\description{
Run all analyses and produce outputs in the \code{outputs_dir} folder
}

